Initial work to move address from being a singleton
By moving address to something that is created before used, we not
only add a configuration api (i.e. construction arguments) but we
also get rid of the internal loading of resources, meaning we no
longer makes the sames assumptions of the host environment.

This is a major breaking change, and the version has been bumped
accordingly.