Fix broken test
We think this test was broken before, we just happened to find while changing
the code it is testing.

We are passing params as an object now, instead of passing an array of keys. The
params are later on interpolated in a path pattern, so from that perspective
it does not make sense to pass an array of keys.