Implement polyfill for element.matches
IE8 doesn't implement element.matches, so we implement a simple
polyfill that will use document.querySelectorAll instead. This is
less efficient than element.matches, but at least it works.
The matches function will always return false in case the node isn't
added to the document, and so we shortcut that case.