Refactor location to not make use of .origin API
The .origin API (of anchor elements and window.location) doesn't
work consistently across browsers, so this refactor should limit
its use to where it's available, and otherwise generate it as
necessary. As well, IE adds a default port value for anchors even
if it's not specified, based on the protocol. However, it's not
included in the href so we need to make sure it's ignored.