Refactored to make hello-world example self-contained
This changes the hello-world example to be a self contained project
complete with its own set of dependencies and scripts. This ensures we
don't pollute the main address project with specifics related to the
examples we come up with, and also makes it easier to make examples that
utilize different stacks – this one uses [http-server] but we might show
another example that integrates with [express], for instance.

[http-server]: https://www.npmjs.com/package/http-server
[express]: https://www.npmjs.com/package/express