update compose to look at metadata for props
These props include:

- `composes` - list of routes that a route composes
- `path` - the pattern or url path associated with the route
- `redirects`- object where keys represent path redirect from and the
values being the path to be redirect to.

Because of the lack of documentation, there are uses of address where a
set of customisation have been made:
+ `web.find` is overridden.
+ metadata properties is stored on the route resource itself, rather
than on `resource.metadata`.

In a future release we may want to deprecate or remove this behaviour.
Further details can be found:
https://github.com/zambezi/address/issues/59